chore: remove any nested index barrel files
