--- packages/core/src/__mocks__/index.ts
@@ -1,1 +1,0 @@
-export * from './output';
