+++ packages/core/src/__tests__/cyclic-package-graph-node.spec.ts
@@ -5,7 +5,7 @@ import { dirname } from 'node:path';
 import { Project } from '../project/project.js';
 import { initFixtureFactory } from '@lerna-test/helpers';
 import { CyclicPackageGraphNode } from '../package-graph/lib/cyclic-package-graph-node.js';
-import { PackageGraphNode } from '../package-graph/index.js';
+import { PackageGraphNode } from '../package-graph/lib/package-graph-node.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
