--- packages/core/src/corepack/index.ts
@@ -1,2 +1,0 @@
-export * from './exec-package-manager.js';
-export * from './is-corepack-enabled.js';
