+++ packages/core/src/index.ts
@@ -1,9 +1,43 @@
 // folders
-export * from './corepack/index.js';
-export * from './models/index.js';
-export * from './package-graph/index.js';
-export * from './project/index.js';
-export * from './utils/index.js';
+export * from './corepack/exec-package-manager.js';
+export * from './corepack/is-corepack-enabled.js';
+export * from './models/command-options.js';
+export * from './models/interfaces.js';
+export * from './package-graph/package-graph.js';
+export * from './package-graph/lib/cyclic-package-graph-node.js';
+export * from './package-graph/lib/package-graph-node.js';
+export * from './package-graph/lib/report-cycles.js';
+export * from './project/project.js';
+export * from './project/lib/apply-extends.js';
+export * from './project/lib/make-file-finder.js';
+export * from './project/lib/shallow-extend.js';
+export * from './utils/collect-updates/collect-updates.js';
+export * from './utils/collect-updates/lib/collect-packages.js';
+export * from './utils/collect-updates/lib/collect-dependents.js';
+export * from './utils/collect-updates/lib/get-packages-for-option.js';
+export * from './utils/collect-updates/lib/has-tags.js';
+export * from './utils/collect-updates/lib/make-diff-predicate.js';
+export * from './utils/check-working-tree.js';
+export * from './utils/clean-stack.js';
+export * from './utils/collect-uncommitted.js';
+export * from './utils/defaults.js';
+export * from './utils/describe-ref.js';
+export * from './utils/env-replace.js';
+export * from './utils/find-prefix.js';
+export * from './utils/log-package-error.js';
+export * from './utils/npm-conf.js';
+export * from './utils/object-utils.js';
+export * from './utils/output.js';
+export * from './utils/parse-field.js';
+export * from './utils/prerelease-id-from-version.js';
+export * from './utils/pulse-till-done.js';
+export * from './utils/query-graph.js';
+export * from './utils/run-lifecycle.js';
+export * from './utils/run-topologically.js';
+export * from './utils/string-utils.js';
+export * from './utils/types.js';
+export * from './utils/warn-if-hanging.js';
+export * from './utils/write-log-file.js';
 
 // files
 export * from './child-process.js';
