--- packages/core/src/models/index.ts
@@ -1,2 +1,0 @@
-export * from './command-options.js';
-export * from './interfaces.js';
