--- packages/core/src/package-graph/index.ts
@@ -1,2 +1,0 @@
-export * from './package-graph.js';
-export * from './lib/index.js';
