--- packages/core/src/package-graph/lib/index.ts
@@ -1,3 +1,0 @@
-export * from './cyclic-package-graph-node.js';
-export * from './package-graph-node.js';
-export * from './report-cycles.js';
