--- packages/core/src/project/index.ts
@@ -1,2 +1,0 @@
-export * from './project.js';
-export * from './lib/index.js';
