--- packages/core/src/project/lib/index.ts
@@ -1,3 +1,0 @@
-export * from './apply-extends.js';
-export * from './make-file-finder.js';
-export * from './shallow-extend.js';
