--- packages/core/src/utils/collect-updates/index.ts
@@ -1,2 +1,0 @@
-export * from './collect-updates.js';
-export * from './lib/index.js';
