+++ packages/core/src/utils/collect-updates/lib/has-tags.ts
@@ -1,7 +1,7 @@
 import { log } from '@lerna-lite/npmlog';
 
 import { execSync } from '../../../child-process.js';
-import { ExecOpts } from '../../../models/index.js';
+import { ExecOpts } from '../../../models/interfaces.js';
 
 /**
  * Determine if any git tags are reachable.
