--- packages/core/src/utils/collect-updates/lib/index.ts
@@ -1,5 +1,0 @@
-export * from './collect-packages.js';
-export * from './collect-dependents.js';
-export * from './get-packages-for-option.js';
-export * from './has-tags.js';
-export * from './make-diff-predicate.js';
