--- packages/core/src/utils/index.ts
@@ -1,22 +1,0 @@
-export * from './collect-updates/index.js';
-export * from './check-working-tree.js';
-export * from './clean-stack.js';
-export * from './collect-uncommitted.js';
-export * from './defaults.js';
-export * from './describe-ref.js';
-export * from './env-replace.js';
-export * from './find-prefix.js';
-export * from './log-package-error.js';
-export * from './npm-conf.js';
-export * from './object-utils.js';
-export * from './output.js';
-export * from './parse-field.js';
-export * from './prerelease-id-from-version.js';
-export * from './pulse-till-done.js';
-export * from './query-graph.js';
-export * from './run-lifecycle.js';
-export * from './run-topologically.js';
-export * from './string-utils.js';
-export * from './types.js';
-export * from './warn-if-hanging.js';
-export * from './write-log-file.js';
