+++ packages/core/src/utils/run-lifecycle.ts
@@ -3,7 +3,7 @@ import PQueue from 'p-queue';
 import { log } from '@lerna-lite/npmlog';
 
 import { npmConf } from '../utils/npm-conf.js';
-import { LifecycleConfig } from '../models/index.js';
+import { LifecycleConfig } from '../models/interfaces.js';
 import { Package } from '../package.js';
 
 const queue = new PQueue({ concurrency: 1 });
