+++ packages/list/src/__tests__/list-command.spec.ts
@@ -10,8 +10,8 @@ vi.mock('@lerna-lite/core', async () => ({
   promptTextInput: (await vi.importActual<any>('../../../core/src/__mocks__/prompt')).promptTextInput,
   throwIfUncommitted: (await vi.importActual<any>('../../../core/src/__mocks__/check-working-tree')).throwIfUncommitted,
   collectUpdates: (await vi.importActual<any>('../../../core/src/__mocks__/collect-updates')).collectUpdates,
-  PackageGraph: (await vi.importActual<any>('../../../core/src/package-graph')).PackageGraph,
-  getPackages: (await vi.importActual<any>('../../../core/src/project')).getPackages,
+  PackageGraph: (await vi.importActual<any>('../../../core/src/package-graph/package-graph')).PackageGraph,
+  getPackages: (await vi.importActual<any>('../../../core/src/project/project')).getPackages,
 }));
 
 vi.mock('@lerna-lite/list', async () => await vi.importActual<any>('../list-command'));
