+++ packages/publish/src/index.ts
@@ -1,3 +1,22 @@
-export * from './lib/index.js';
+export * from './lib/create-temp-licenses.js';
+export * from './lib/fetch-config.js';
+export * from './lib/get-current-sha.js';
+export * from './lib/get-current-tags.js';
+export * from './lib/get-npm-username.js';
+export * from './lib/get-packages-without-license.js';
+export * from './lib/get-packed.js';
+export * from './lib/get-profile-data.js';
+export * from './lib/get-tagged-packages.js';
+export * from './lib/get-two-factor-auth-required.js';
+export * from './lib/get-unpublished-packages.js';
+export * from './lib/get-whoami.js';
+export * from './lib/git-checkout.js';
+export * from './lib/log-packed.js';
+export * from './lib/npm-dist-tag.js';
+export * from './lib/npm-publish.js';
+export * from './lib/pack-directory.js';
+export * from './lib/remove-temp-licenses.js';
+export * from './lib/throttle-queue.js';
+export * from './lib/verify-npm-package-access.js';
 export * from './interfaces.js';
 export * from './publish-command.js';
