--- packages/publish/src/lib/index.ts
@@ -1,20 +1,0 @@
-export * from './create-temp-licenses.js';
-export * from './fetch-config.js';
-export * from './get-current-sha.js';
-export * from './get-current-tags.js';
-export * from './get-npm-username.js';
-export * from './get-packages-without-license.js';
-export * from './get-packed.js';
-export * from './get-profile-data.js';
-export * from './get-tagged-packages.js';
-export * from './get-two-factor-auth-required.js';
-export * from './get-unpublished-packages.js';
-export * from './get-whoami.js';
-export * from './git-checkout.js';
-export * from './log-packed.js';
-export * from './npm-dist-tag.js';
-export * from './npm-publish.js';
-export * from './pack-directory.js';
-export * from './remove-temp-licenses.js';
-export * from './throttle-queue.js';
-export * from './verify-npm-package-access.js';
