+++ packages/run/src/index.ts
@@ -1,1 +1,4 @@
+export * from './lib/get-npm-exec-opts.js';
+export * from './lib/npm-run-script.js';
+export * from './lib/timer.js';
 export * from './run-command.js';
