+++ packages/run/src/lib/__tests__/npm-run-script.spec.ts
@@ -4,7 +4,7 @@ import { describe, expect, it, Mock, vi } from 'vitest';
 vi.mock('@lerna-lite/core');
 
 import { execPackageManager, spawnStreaming } from '@lerna-lite/core';
-import { RunScriptOption, ScriptStreamingOption } from '../../models/index.js';
+import { RunScriptOption, ScriptStreamingOption } from '../../interfaces.js';
 
 // file under test
 import { npmRunScript, npmRunScriptStreaming } from '../npm-run-script.js';
