--- packages/run/src/lib/index.ts
@@ -1,3 +1,0 @@
-export * from './get-npm-exec-opts.js';
-export * from './npm-run-script.js';
-export * from './timer.js';
