+++ packages/version/src/__mocks__/conventional-commits/apply-build-metadata.ts
@@ -1,0 +1,12 @@
+import { vi } from 'vitest';
+
+const mockApplyBuildMetadata = vi.fn().mockName('applyBuildMetadata');
+
+mockApplyBuildMetadata.mockImplementation((version, buildMetadata) => {
+  if (buildMetadata) {
+    return `${version}+${buildMetadata}`;
+  }
+  return version;
+});
+
+export const applyBuildMetadata = mockApplyBuildMetadata;
