+++ packages/version/src/__mocks__/conventional-commits/recommend-version.ts
@@ -1,0 +1,8 @@
+import { vi } from 'vitest';
+import semver from 'semver';
+
+const mockRecommendVersion = vi.fn().mockName('recommendVersion');
+
+mockRecommendVersion.mockImplementation((node) => semver.inc(node.version, 'patch'));
+
+export const recommendVersion = mockRecommendVersion;
