+++ packages/version/src/__mocks__/conventional-commits/update-changelog.ts
@@ -1,20 +1,8 @@
 import { vi } from 'vitest';
 import { join } from 'node:path';
-import semver from 'semver';
 const { outputFile } = await vi.importActual<any>('fs-extra/esm');
 
-const mockApplyBuildMetadata = vi.fn().mockName('applyBuildMetadata');
-const mockRecommendVersion = vi.fn().mockName('recommendVersion');
 const mockUpdateChangelog = vi.fn().mockName('updateChangelog');
-
-mockApplyBuildMetadata.mockImplementation((version, buildMetadata) => {
-  if (buildMetadata) {
-    return `${version}+${buildMetadata}`;
-  }
-  return version;
-});
-
-mockRecommendVersion.mockImplementation((node) => semver.inc(node.version, 'patch'));
 
 mockUpdateChangelog.mockImplementation((pkg) => {
   const filePath = join(pkg.location, 'CHANGELOG.md');
@@ -26,6 +14,4 @@ mockUpdateChangelog.mockImplementation((pkg) => {
   }));
 });
 
-export const applyBuildMetadata = mockApplyBuildMetadata;
-export const recommendVersion = mockRecommendVersion;
 export const updateChangelog = mockUpdateChangelog;
