+++ packages/version/src/conventional-commits/get-github-commits.ts
@@ -2,7 +2,7 @@ import { ExecOpts, getComplexObjectValue } from '@lerna-lite/core';
 import { log } from '@lerna-lite/npmlog';
 import dedent from 'dedent';
 
-import { createGitHubClient, parseGitRepo } from '../git-clients/index.js';
+import { createGitHubClient, parseGitRepo } from '../git-clients/github-client.js';
 import { RemoteCommit } from '../interfaces.js';
 
 const QUERY_PAGE_SIZE = 100; // GitHub API is restricting max of 100 per query
