+++ packages/version/src/git-clients/__tests__/gitlab-client.spec.ts
@@ -3,7 +3,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 vi.mock('../GitLabClient');
 
 import { GitLabClient } from '../GitLabClient.js';
-import { createGitLabClient } from '../index.js';
+import { createGitLabClient } from '../gitlab-client.js';
 
 describe('createGitLabClient', () => {
   const oldEnv = Object.assign({}, process.env);
