--- packages/version/src/git-clients/index.ts
@@ -1,2 +1,0 @@
-export * from './github-client.js';
-export * from './gitlab-client.js';
