+++ packages/version/src/index.ts
@@ -1,6 +1,28 @@
-export * from './conventional-commits/index.js';
-export * from './git-clients/index.js';
-export * from './lib/index.js';
+export * from './conventional-commits/apply-build-metadata.js';
+export * from './conventional-commits/constants.js';
+export * from './conventional-commits/get-changelog-config.js';
+export * from './conventional-commits/get-commits-since-last-release.js';
+export * from './conventional-commits/get-github-commits.js';
+export * from './conventional-commits/make-bump-only-filter.js';
+export * from './conventional-commits/read-existing-changelog.js';
+export * from './conventional-commits/recommend-version.js';
+export * from './conventional-commits/update-changelog.js';
+export * from './conventional-commits/writer-opts-transform.js';
+export * from './git-clients/github-client.js';
+export * from './git-clients/gitlab-client.js';
+export * from './lib/create-release.js';
+export * from './lib/get-current-branch.js';
+export * from './lib/git-add.js';
+export * from './lib/git-commit.js';
+export * from './lib/git-push.js';
+export * from './lib/git-tag.js';
+export * from './lib/is-anything-committed.js';
+export * from './lib/is-behind-upstream.js';
+export * from './lib/is-breaking-change.js';
+export * from './lib/prompt-version.js';
+export * from './lib/remote-branch-exists.js';
+export * from './lib/update-lockfile-version.js';
 export * from './interfaces.js';
 export * from './version-command.js';
-export * from './utils/index.js';
+export * from './utils/otplease.js';
+export * from './utils/temp-write.js';
