--- packages/version/src/lib/index.ts
@@ -1,12 +1,0 @@
-export * from './create-release.js';
-export * from './get-current-branch.js';
-export * from './git-add.js';
-export * from './git-commit.js';
-export * from './git-push.js';
-export * from './git-tag.js';
-export * from './is-anything-committed.js';
-export * from './is-behind-upstream.js';
-export * from './is-breaking-change.js';
-export * from './prompt-version.js';
-export * from './remote-branch-exists.js';
-export * from './update-lockfile-version.js';
