--- packages/version/src/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './otplease.js';
-export * from './temp-write.js';
