+++ packages/version/src/version-command.ts
@@ -48,12 +48,10 @@ import {
   saveUpdatedLockJsonFile,
 } from './lib/update-lockfile-version.js';
 import { GitCreateReleaseClientOutput, ReleaseNote, RemoteCommit } from './interfaces.js';
-import {
-  applyBuildMetadata,
-  getCommitsSinceLastRelease,
-  recommendVersion,
-  updateChangelog,
-} from './conventional-commits/index.js';
+import { applyBuildMetadata } from './conventional-commits/apply-build-metadata.js';
+import { getCommitsSinceLastRelease } from './conventional-commits/get-commits-since-last-release.js';
+import { recommendVersion } from './conventional-commits/recommend-version.js';
+import { updateChangelog } from './conventional-commits/update-changelog.js';
 
 export function factory(argv: VersionCommandOption) {
   return new VersionCommand(argv);
